pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "nyc_crashes_dagster_dbt"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "dagster",
    "dagster-webserver",
    "psycopg2-binary",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["nyc_crashes_dagster_dbt*"]

[tool.dagster]
module_name = "nyc_crashes_dagster_dbt.definitions"
